fix(x64): resolve std::map/pair type mismatches and shared str_is_number
- char_skill.cpp: cast size_t loop index to int for std::map<int,LPEVENT>
  key (matches the existing casts at the find/erase sites above it).
- questlua_dungeon.cpp: explicit lua_Number->DWORD/int casts building
  std::pair<DWORD,int> from lua_tonumber (lua numbers are doubles).
- guild_manager.cpp: cast bet gold to DWORD for the mapBet
  pair<DWORD,DWORD> value (guild-war bet storage is 32-bit, same
  contract as item_award.socket0).
- common/utils.h: cast strlen result in str_is_number (bounded).

Server x64 now compiles with only 3 warnings remaining, all deliberate
decisions pending review (switchbot BYTE packet field x2; qc == bug).

+++ Lead-Server-Source/db/src/ClientManager.cpp
@@ -1725,7 +1725,7 @@ void CClientManager::QUERY_AUTH_LOGIN(CPeer * pkPeer, DWORD dwHandle, TPacketGDA
 
 void CClientManager::GuildWarBet(TPacketGDGuildWarBet * p)
 {
-	CGuildManager::instance().Bet(p->dwWarID, p->szLogin, p->dwGold, p->dwGuild);
+	CGuildManager::instance().Bet(p->dwWarID, p->szLogin, static_cast<DWORD>(p->dwGold), p->dwGuild);
 }
 
 void CClientManager::CreateObject(TPacketGDCreateObject * p)
